requirements.txt
certifi==2019.11.28
cffi==1.13.2
Click==7.0
cryptography==2.8
dlib==19.19.0
face-recognition==1.2.3
face-recognition-models==0.3.0
future==0.18.2
gimpify-watxaut==0.0.1a3
numpy==1.18.1
Pillow==7.0.0
pycparser==2.19
python-telegram-bot==12.2.0
six==1.13.0
tornado==6.0.3
